Move index directories to separate module
This allows multiple modules to import the index directory codecs.
